Move player and proxy depot classes to separate packages
