+++ proxy/src/main/java/com/velocityctd/proxy/cluster/redis/RedisClusterPlayerService.java
@@ -21,8 +21,8 @@ import com.velocityctd.proxy.cluster.VelocityClusterPlayer;
 import com.velocityctd.proxy.cluster.VelocityClusterPlayerService;
 import com.velocityctd.proxy.redis.VelocityRedis;
 import com.velocityctd.proxy.redis.data.VelocityAlert;
-import com.velocityctd.proxy.redis.depot.PlayerDepotService;
-import com.velocityctd.proxy.redis.depot.PlayerEntry;
+import com.velocityctd.proxy.redis.depot.player.PlayerDepotService;
+import com.velocityctd.proxy.redis.depot.player.PlayerEntry;
 import com.velocitypowered.api.proxy.player.PlayerSettings;
 import com.velocitypowered.proxy.VelocityServer;
 import com.velocitypowered.proxy.connection.client.ConnectedPlayer;
